Package: @encapsule/arccore v0.1.7 "halfmoon" build ID "OEaB2oLMTlao80edIE5wlQ" Sources: Encapsule/ARC_master#3a38e1b32e55d9a7da60c6307217f8b45927c088
This release patches the first bug in arccore.filter discovered in nearly eighteen months of daily production use.

The documentation states that a filter's response.result is only valid per the filter's output filter spec iff ~!response.error (response.error === null)

And, if response.error (response.error !== null in which case it is always a string) then response.result is supposed to be left as it was returned by bodyFunction allowing bodyFunction to use response.result as required for error reporting (i.e. it's opaque to filter iff response.error).

v0.1.6 grasslands and earlier @encapsule/arccore package releases would incorrectly set response.result = null if bodyFunction returned a well-formed response with a result that failed the output filter spec validation.

Now, we leave it alone in this case. Note that in cases where there's no output filter spec, then this is the behavior you've always gotten from just the plain bodyFunction (because we just check this case for a well-formed reponse and don't worry about determining anything about response.result).